ScooterBackend/pyproject.toml
[tool.poetry]
name = "scooter24"
version = "0.1.0"
description = ""
authors = ["darkfos82 <darkfos82@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.111.0"}
SQLAlchemy = "^2.0.30"
asyncpg = "^0.29.0"
black = "^24.4.2"
alembic = "^1.13.1"
python-jose = "^3.3.0"
PyJWT = "^2.8.0"
passlib = "^1.7.4"
aiohttp = "^3.9.5"
pillow = "^10.3.0"
requests = "^2.32.3"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
sqladmin = "^0.19.0"
redis = "^5.0.8"
celery = "^5.4.0"
flower = "^2.0.1"
flake8 = "^7.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
